ref(ui): Extract EDR blending logic into EDRMetalImageView
Move the CIImage scaling and headroom blending out of QRCodeImageView
into a dedicated EDRMetalImageView component for better separation of
concerns.
